# Simplify unprintable handling in UnpackDomainName