--- a/msg.go
+++ b/msg.go
@@ -405,7 +405,7 @@ Loop:
 				case '"', '\\':
 					s = append(s, '\\', b)
 				default:
-					if b < 32 || b >= 127 { // unprintable, use \DDD
+					if b < ' ' || b > '~' { // unprintable, use \DDD
 						var buf [3]byte
 						bufs := strconv.AppendInt(buf[:0], int64(b), 10)
 						s = append(s, '\\', '0', '0', '0')